Update svgExport.pspscript
Added new version of the group conversion function to be used instead of the old.  This one converts all vector objects in the group to paths before continuing with reading the group data.  As there is no symmetric shape option in the SVG format, and PSP doesn't return all Text data to faithfully recreate text in SVG format.